Fixes a minor client disconnection runtime (#60834)
This happened because clients are incredibly fickle and may disappear at any time.
In this case searching for the DB took long enough to cause this.
Improved a little code around it, in the process.

(It's more like the db sleeps, so byond has a chance to nuke the client datum. but this is good either way)