Revival surgery works with shock touch (#61329)
Just add shock touch as an alternative to the defib paddles on the revival surgery, just like you can use a stun baton.
To be clear, it doesn't work as a defib for normal revival, just as a "ghetto" alternative to revival surgery.

Also I'm putting this as a QoL since it is something small on the grand scheme of Genetics, if a maintainer things that a Balance/Content expansion tag would be more fitting, feel free to add it.
As a side note, shock touch has 30 of instability, the same as Space Adapt.