Reorder HFR fuel datums (#61297)
* Reorder HFR fuel datums

This reorders the HFR datums in order of maximum temperature. This is
also a rough approximation of difficulty and order of desirability.

This mostly stops the plasma/oxygen recipe being a weird outlier in the
middle of the interface, but also somewhat makes the recipes above
make more sense too.

No functional changes.