Actually fixes temperature pump exploit. Removes limit as it's redundant now. (#61406)
Temperature pumps were bugged and generated heat while exchanging heat. This fixes that and removes the limits as they were only used to stop people from creating infinite heat.

Exploit bad, and the limit was below a temperature that can be obtained legit(imately, which was annoying.

(I was initially against this pr, was under the impression that the heat pump actually pumped heat. apparently it doesn't. currently considering if it even needs to exist.)