Pushing things by bumping into them now adds fingerprints. (#61507)
Bumping into things to push them didn't add fingerprints. This means that it's possible to push things like fuel tanks without any way to track IC or OOC.

Now when you bump into a movable atom and it successfully gets moved by you, fingerprints get added the same as if you'd grabbed it and pulled it.