Lizard Firebreath is now a cone of fire instead of a fireball (#61742)
This PR reworks the firebreath mutation from being a wizard fireball that one hit crits with a power chromosome to a wave of fire.
The cone is 3 tiles, or 5 with the power chromosome, and travels in mostly a three wide column.
The closer you are to the caster, the more fire damage you take.
Non-antags can get wizard fireball and that's really dumb, this re-flavors it to be more like a dragon's breath, while removing the stupid aspect.