Update swarmer_objs.dm (#61877)
If you are banned from an antag, but no ghosts take your body, you get to stay in your body.
This means that someone banned from Swarmer can keep trying to join as a Swarmer and hope no one will take over their body, to bypass their ban.
To prevent this, banned people won't be able to form a swarmer body in the first place.