Fix CentCom areas/APCs (#61964)
This adds dedicated areas for places that have APCs. It also adds an APC
to the ferry Administrative Office, as it seems unusual to only have one
in Administrative Storage, which has distinct (ACCESS_CENT_CAPTAIN!)
access requirements.

APCs are still responsible for multiple centcom areas elsewhere, as it
seems a bit wasteful to have them everywhere. But this should keep things
coherent if anything happens on CentCom where access distinctions
(or at least, captain tier access!) matters.

Names of the new `/area`s are taken from the names on the airlocks that
lead to them.