Fix antagonist tooltip showing wrong description (#61973)
This fixes the TGUI where users hovers over antagonists in the preference menu. The problem was that when moving your mouse over multiple antags, their descriptions would be shown for other antags.

Before video:
https://user-images.githubusercontent.com/30943236/136470845-3ec9bbf0-0453-4658-a895-4f8e89a8f5cb.mp4

After video:
https://user-images.githubusercontent.com/30943236/136470888-49ef7a77-f142-47c1-8178-78add8db906b.mp4

This was caused because the descriptions are an array that is iterated on and the key used for each description item is the index (0 or 1). The fix was to create a unique key, this was done by concatenating the antagonist key (example, Wizard is 'wizard') with the index (so wizard0, wizard1).