Fixes a very small mistake with the clickon signal procs cleanup. (#62214)
I haven't realized that the target loc check on clowncar.dm would have prevented players from firing the emag mode cannon when clicking on turfs (because their loc is an area). So I'm adding a ismovable(target) check before the target.loc in the same conditional statement.
Also renaming the `A` argument to `target`.