Reorganizes the order of items in job outfits (#61553)
Outfits didn't have an order to the list of items. I started organizing them to be easier to read and find mistakes on #57385 and now I got to job outfits.

The order is based on "priority" as follows:
ID > Uniform > Suit > Backpack > Other clothing in alphabetical order > Pockets > Hands > Others