Pressure Tank: Fix many runtimes, consistency (#62317)
Use MINERAL_MATERIAL_AMOUNT, add and use TANK_PLATING_SHEETS in place of magic numbers everywhere.

If you get a spooky warning about the tank being pressurized as you slice it open with a welding tool, you actually get consequences. These consequences (and the new conditions for it) are consistent with the checks in atmosmachinery.dm.

Don't try to use an array index of 2 on the custom_materials associative list which is meant to hold a single entry. If custom_materials ends up with multiple materials somehow, produce a frame with nothing rather than a plated frame.

Properly pass the amount of a material to create when dropping.

Don't try to use a stackable item which doesn't have a material_type.