Fixes exploit that allowed players to disable the safety of holodeck consoles without the proper conditions being followed (#62563)
About The Pull Request

As the title says. Players were able to disable the safety of the holodeck without the console being emagged, and without being a robot.
Why It's Good For The Game

I think it would be interesting to add some sort of replacement to this, like players with high gamer skill could disable the safety.
Changelog

cl
fix: Fixes an exploit that allowed players to disable holodeck safety without following the proper conditions
/cl