The Nuke kills everyone when it goes off. (#62807)
This does not make the nuke explode like it used to.
The nuke will now dust you in a similar way to the AI doomsday device when it goes off. This will only happen if the nuke successfully detonated on the station and not on a near miss. Otherwise, it kills everything on the z-level.

There has been code in the game for an extended period of time that was supposed to make this happen, but I believe it was bugged and did not properly give the z-level of the nuke in the explosion.

It should be noted that gameplay-wise this makes it so antagonists who need to escape will no longer get a free pass for being on the station when the nuke goes off, but they will still be counted as escaped if they are in lavaland/another z-level. I would say that this is only good for the game but I understand if it's controversial.

The original code gibbed people on the z-level, although I've considered changing it to dusting. I would like opinions on this or alternative methods of death.

fixes #62788