Makes hypnosis victims more visible to admins (#62745)
Gives hypnotized victims a visible antag hud for admins, puts them in the Check Antags panel, and makes their objectives visible in the Traitor Panel and on the round end report.
Why It's Good For The Game

Right now, victims of hypnosis are very difficult to visually identify as an admin. This not only distinguishes them easily but makes their objective easily visible too.
Changelog

cl
admin: hypnotized players now have a visible antag hud, are visible in the check antag panel, and have objectives visible on the traitor panel
expansion: hypnotized victims and their objectives are now visible on the round end report
/cl