Fixes pipe layer and rotation actions interfering (#63059)
Currently, attempting to left + alt click a pipe or pipe accessory will both rotate and change the layer the pipe sits on in the same click. This changes the layer swapping to occur on right + alt click while leaving the rotation action to a left + alt click.

Fixes #62809 (and improves atmospherics game-play):
    try and rotate a pipe with alt click
    it changes its layer too
    be sad