performance improvement for process_cell() (#62730)
About The Pull Request

Increase performances on /turf/open/process_cell() by changing a typed for and making use of as anything, plus removed unnecessary proc call
Why It's Good For The Game

The test has been done in a controlled environment, a 7x7 airless room with one canister of plasma and one canister of oxygen released inside. The profiling ran from the start of the fire to the end of it.
Before any changes:
image
After the as anything forloop:
image
After the archive() removal:
image
Changelog

cl
code: increased atmos performances a bit more
/cl