Fix sec hailer oversight with tint changes (#63211)
About The Pull Request

Sec hailers should've been changed to not have a tint but also to not have eye protection and so pepper immunity.
Why It's Good For The Game

Fix an oversight on the original PR #62947
Changelog

cl
fix: fix an oversight where sec hailer shouldn't have the tint changes, the eye protection and the pepper immunity. All other sec masks (the swat and spacepol) are unchanged
/cl