Fix AI eye not being recreated if it was qdel'd, but not nulled (#63131)
As an error recovery point, if the eye somehow gets qdel'd, it is recreated when you try to move to the core.

However, this error recovery doesn't work if the AI eye is qdel'd without being nulled, as create_eye() checks if eyeobj exists (NOT if it's qdel'd).

Of course the real bug is in whatever is qdel'ing the eye, but considering this is an existing error recovery system, it should still work!