Makes the edible component not runtime due to connect_loc_behalf when applied to turfs (#63232)
This makes it so it stops trying to use connect_loc_behalf if the parent of the edible component is a turf, just registering the signal normally instead, which should hopefully be the same functionality.

Runtimes are bad, and even if there's not really any relevant use of that proc in most cases, this is to future-proof things anyway.