returns functionality to machine sect's self android conversion (#63215)
Chaplains can now invoke the rite with nobody on the altar to convert themselves. The bugged functionality of religious tool that let anyone invoke rites allowed chaplains to be converted by others, which became somewhat of a core feature of the sect. In fixing the previous bug the sect lost a somewhat large part of itself, and I think the best solution is to just make it intended

(cherry picked from commit 12f71adea2b2b857f5479dbdb0893aca4cb672d7)