Increases volume of the sound effect for knocking on firelocks (#63053)
Changed firelocks to have the same bash and knock sound effects as windows.
Which increases the volume of the knock sound effect and also adds the bash sound effect.

Also improves some of the code for firelock and window

Firelocks already had the knock sound effect, but it is so quiet that you can barely hear it. The volume should be increased so that people can hear it.

(cherry picked from commit 2ee3d602fc232f2e9fef01d761c44aa9563d96c2)