Fixes bad taste description logic for fruit wine (#63421)
OpenDream sussed out a bug with switch cases and wixoa confirmed that it's actually a BYOND compiler bug:
BYOND isn't handling it properly at all (it should throw a "expected constant expression" error), but converting the var to a const fixes it. Since, y'know, switch cases need to be consts.