Fixes check on Rat King's rat spit OD effect not checking if victim is already in faction. (#63426)
The regal rat has an underutilized mechanic where its rat spit reagent causes victim that hit the OD threshold to be added to the rat faction along with a nice big red flavor text. That proc lacked a check to see if the victim was already within the faction which this PR aims to fix.
Prior to this PR there existed no check to see if the victim was already within the faction, causing it to proc every other time the OD threshold was hit and giving the player the flavor text about how they had just obtained a new leader despite it already having been the case. Also avoids trying to add the victim to the faction they already were in.

(cherry picked from commit 91d1f4cc52a7e512f48a83089cae05b775a1e3b5)