Add accessibility label--same points as a fix (#63593)
Adds an Accessibility label. This label should be added for any feature that improves the experience for people with disabilities.

This includes, but is not limited to:

    Adding visual/auditory feedback for which there previously was none (such as adding a balloon alert where the only other feedback was a sound)
    Adding captions, such as to AI vox, ambience, etc
    Adding colorblind options, or otherwise making features not rely on colors (such as pipes, slimes, etc)
    Adding more keyboard navigation to UIs, such that a mouse is not as required
    Reducing flashing/bright lights for those with epilepsy

This does not include:

    General QoL, such as replacing chat messages with balloon alerts
    Making the game generally easier to do weird interactions with. This is considered bettering approachability, not necessarily accessibility

...though many changes that improve QoL can easily improve accessibility.

For further reading, the Game Accessibility Guidelines are a very interesting resource to go through.

(cherry picked from commit 4e80c486fe1236e0e9ffc24d80b7f6a0064e8016)