Optimizes the `output_part_info()` proc in Mech Fabricators. (#63379)
It turns out the mech fabricator updated the list of what was available every time anyone opened it, which is incredibly expensive.

I optimized it to only update the list on Init and when the console is re-synced with the R&D console, in addition to making the proc much, much lighter to run.

(cherry picked from commit 7d23b40526f8e694d14158fde8242ea1e4dabf38)