Fix supermatter and door runtimes (#63718)
About The Pull Request

It turns out that when the supermatter is about to explode it spams a div/0 runtime instead of emitting rads. This fixes that.
This also fixes a runtime caused by attempting to walk through an airlock with a cut ID wire since it came up in testing.
Why It's Good For The Game

Less runtimes.
Changelog

cl
fix: The supermatter will once again emit radiation when it is about to explode.
/cl

(cherry picked from commit 5e8edc246ce3e5676428a2a799e1299243702a99)