Writes out lists for roles in the Dynamic Rulesets Code (#63970)
Alphabetizes and spans out the lists for all of the restricted_roles and protected_roles in the three different DM files for the Dynamic rulesets. @Timberpoes in #63960 already addresses this for Latejoin Provocateur and Roundstart Revolutionary and Cultist, so those are untouched.

(cherry picked from commit 411693973ec936f8a14426d3ec043efc8508ba34)