Fixes cyborg light overlay not showing when the light is off (#64279)
Borgs have a lightbulb overlay. The plane of the overlay depends on the status of the lights, so as to glow when on. The plane being used for disabled lights no longer shows above mobs for reasons I do not know. But this plane (-2) works.

(cherry picked from commit b6f32d464730c8e7e65691adb4f073a6efdb3cc7)