Add's food processor function to spider eggs and carp filet. Yaki imo recipe change (#64704)
At the moment carp meat kind of has no use at all for the chef, the meat is poisonous which is just kind of a drag. The only practical use for carp at the moment is rezadone.

Also uh, we have ovens now so yaki imo should be made in the oven.