Fixes the void Ritual of Knowledge + unit tests main path lengths  (#64929)
Void ritual of knowledge should be of the void path. This fixes that.

Also adds a unit test to ensure all main paths are of the same length.

Funnily I think this was causing the main route calculations to be off, so the knowledge objective should be easier by one.