Replacing power cells with RPED takes their charge into account (#65033)
Instead of get_part_ratings() (which is what the RPED uses to decide whether one stock part is better than another) returning rating * maxcharge on cells, it now returns maxcharge + charge. This means the cell of the highest capacity with the most charge will go in first.

Removes manually set ratings on cells in favour of calculating them in Initialize() in a more consistent and logical way. I don't think it's used for anything now anyway.