Fixes RCD not being able to be put in containers by click (#65249)
If you click a bag with an RCD nothing happends, this fixes it. Not sure if this is the best way to fix it since im not that experienced but i playtested it and it works. Tell me if i need to change anything.

(cherry picked from commit dcb1e2ed2eb3374db99311057f58e7c77d48e1c8)