Fix jumpsuits to drop accessories when destroyed (#65768)
Fixes #65103 and probably some other edge cases.

Medals of captaincy are an indestructible antag objective. If you attached the medal to the jumpsuit and somehow destroyed it, it would delete all the accessories on it. Now when jumpsuits are destroyed, they drop their accessories to the floor.

(cherry picked from commit bdd0192168d67763e608c67e75308daf22cd285d)